Move ChartModel classes to components.runtime package; Add generic parameter for ChartModel
