+++ appinventor/components/src/com/google/appinventor/components/runtime/ChartDataBase.java
@@ -1,13 +1,10 @@
 package com.google.appinventor.components.runtime;
 
-import com.github.mikephil.charting.charts.Chart;
 import com.google.appinventor.components.annotations.DesignerProperty;
 import com.google.appinventor.components.annotations.PropertyCategory;
 import com.google.appinventor.components.annotations.SimpleObject;
 import com.google.appinventor.components.annotations.SimpleProperty;
 import com.google.appinventor.components.common.PropertyTypeConstants;
-import com.google.appinventor.components.runtime.util.ChartModel;
-import com.google.appinventor.components.runtime.util.ElementsUtil;
 
 @SimpleObject
 public abstract class ChartDataBase implements Component {
