+++ appinventor/components/src/com/google/appinventor/components/runtime/LineChart.java
@@ -1,15 +1,9 @@
 package com.google.appinventor.components.runtime;
 
 import android.view.View;
-import com.github.mikephil.charting.data.Entry;
 import com.github.mikephil.charting.data.LineData;
-import com.github.mikephil.charting.data.LineDataSet;
 import com.google.appinventor.components.annotations.*;
 import com.google.appinventor.components.common.ComponentCategory;
-import com.google.appinventor.components.runtime.util.ChartModel;
-import com.google.appinventor.components.runtime.util.LineChartModel;
-
-import java.util.ArrayList;
 
 @DesignerComponent(version = 1,
     category = ComponentCategory.CHARTS,
